m_swap: use fixed-width types for the byte-swap macros
doom_swap_l cast its result through (long int), which is 64-bit on LP64
targets, while the swap arithmetic only ever handles 32 bits -- so LONG()
and the doom_*l conversion macros yielded a 64-bit type for what is always
32-bit WAD/network data. Switch the swap math and the little-endian identity
casts to int32_t/int16_t (with stdint.h pulled in) so the macro result type
matches the data width on every target.

This is purely a type-correctness fix: the macros remain compile-time only
with no function calls, the swapped values are unchanged (verified bit-
identical across zero/positive/negative/sign-bit/boundary inputs for both
widths), and on little-endian the identity casts generate the same code as
before (single mov / movzwl).

Also drop the now-doubly-redundant (signed short) cast on the GENMIDI
base_note_offset add: the field is already short and doom_htows now yields
int16_t.

+++ src/oplplayer.c
@@ -727,7 +727,7 @@ static unsigned int FrequencyForVoice(opl_voice_t *voice)
 
     if ((voice->current_instr->flags & GENMIDI_FLAG_FIXED) == 0)
     {
-        note += (signed short) doom_htows(gm_voice->base_note_offset);
+        note += doom_htows(gm_voice->base_note_offset);
     }
 
     // Avoid possible overflow due to base note offset:
